moved change-prefix into misc
